chore: enable webpack builds for proxy
part of #8219

+++ packages/webpack/bin/watch.sh
@@ -50,14 +50,18 @@ else
     rm -f $LOCKFILE
 fi
 
-LOCKFILE=$LOCKFILE npx --no-install webpack serve --progress --config "$CONFIG" $OPEN &
-
 if [ -n "$KUI_HEADLESS_WEBPACK" ]; then
-    LOCKFILE2=/tmp/kui-build-lock2.${PORT_OFFSET-0}
-    rm -f $LOCKFILE2
+    echo "Watching Kui Headless bundles via webpack"
+    if [ -n "$LOCKFILE" ]; then
+        LOCKFILE2=/tmp/kui-build-lock2.${PORT_OFFSET-0}
+        rm -f $LOCKFILE2
+    fi
 
     LOCKFILE=$LOCKFILE2 npx --no-install webpack watch --progress --config "$HEADLESS_CONFIG" &
 fi
+
+echo "Watching Kui Client bundles via webpack"
+LOCKFILE=$LOCKFILE npx --no-install webpack serve --progress --config "$CONFIG" $OPEN &
 
 if [ -n "$LOCKFILE" ]; then
     # don't exit until the dev server is ready
